requirements.txt
fpdf==1.7.2
requests==2.32.5
python-dotenv
google-genai
google-cloud-modelarmor